workshop/rest_api/requirements.txt
fastapi
uvicorn 
pydantic 
mysql-connector-python